+++ R/mat_to_net.R
@@ -1,22 +1,28 @@
 #' Create a data.frame from a contingency table
 #'
 #' This function creates a two- or three-columns `data.frame` where
-#' each row represents the interaction between two nodes (site and species for 
+#' each row represents the interaction between two nodes (site and species for
 #' example) and an optional third column indicating the weight of the
 #' interaction (if `weight = TRUE`) from a contingency table (sites as
 #' rows and species as columns for example).
 #'
 #' @param mat a contingency table (i.e. `matrix`).
-#' 
+#'
 #' @param weight a `boolean` indicating if the value are weights.
-#' 
-#' @param remove_zeroes a `boolean` determining whether interactions with 
+#'
+#' @param remove_zeroes a `boolean` determining whether interactions with
 #' weight equal to 0 should be removed from the output.
-#' 
+#'
+#' @param include_diag a `boolean` indicating whether the diagonal
+#' should be included in the output. Only for squared matrix.
+#'
+#' @param include_lower a `boolean` indicating whether the lower triangular
+#' matrix should be included in the output. Only for squared matrix.
+#'
 #' @return A `data.frame` where each row represents the interaction
 #' between two nodes and an optional third column indicating the weight of the
 #' interaction.
-#' 
+#'
 #' @author
 #' Maxime Lenormand (\email{maxime.lenormand@inrae.fr}),
 #' Pierre Denelle (\email{pierre.denelle@gmail.com}) and
@@ -29,37 +35,61 @@
 #' colnames(mat) <- paste0("Species", 1:10)
 #'
 #' net <- mat_to_net(mat, weight = TRUE)
-#' 
+#'
 #' @importFrom tidyr pivot_longer
-#' 
+#'
 #' @export
 
-mat_to_net <- function(mat, weight = FALSE, remove_zeroes = TRUE){
-
+mat_to_net <- function(mat,
+                       weight = FALSE,
+                       remove_zeroes = TRUE,
+                       include_diag = TRUE,
+                       include_lower = TRUE) {
   # Control input mat
   controls(args = NULL, data = mat, type = "input_matrix")
-  
+
   # Control parameters
   controls(args = weight, data = NULL, type = "boolean")
   controls(args = remove_zeroes, data = NULL, type = "boolean")
+  controls(args = include_diag, data = NULL, type = "boolean")
+  controls(args = include_lower, data = NULL, type = "boolean")
+
+  # Squared matrix
+  if (dim(mat)[1] == dim(mat)[2]) {
+    if (!include_diag) {
+      diag(mat) <- 0
+    }
+    if (!include_lower) {
+      mat[lower.tri(mat)] <- 0
+    }
+  } else {
+    if (include_diag == FALSE) {
+      message("include_diag is only used with squared matrix.")
+    }
+    if (include_lower == FALSE) {
+      message("include_lower is only used with squared matrix.")
+    }
+  }
 
   # Visible binding for global variable
   Node1 <- NULL
-  
+
   # Conversion as data.frame
   mat <- as.data.frame(mat)
   mat$Node1 <- rownames(mat)
-  net <- as.data.frame(tidyr::pivot_longer(data = as.data.frame(mat),
-                                           cols = -Node1, names_to = "Node2",
-                                           values_to = "Weight"))
+  net <- as.data.frame(tidyr::pivot_longer(
+    data = as.data.frame(mat),
+    cols = -Node1, names_to = "Node2",
+    values_to = "Weight"
+  ))
 
   # Remove interactions with weight equal 0
-  if(remove_zeroes == TRUE){
+  if (remove_zeroes == TRUE) {
     net <- net[net$Weight != 0, ]
   }
 
   # Remove the weight column if weight is set to FALSE
-  if(!weight){
+  if (!weight) {
     net <- net[, -3]
   }