+++ tests/testthat/test-net_to_mat.R
@@ -5,6 +5,12 @@ net <- data.frame(
   Weight = c(10, 100, 1, 20, 50, 10, 20)
 )
 
+net2 <- data.frame(
+  Site = c(rep("A", 2), rep("B", 3), rep("C", 2)),
+  Species = c("a", "b", "a", "c", "d", "b", "d"),
+  Weight = c("a", "b", "a", "c", "d", "b", "d")
+)
+
 # Tests for valid outputs -----------------------------------------------------
 test_that("dimension of output", {
   mat <- net_to_mat(net, weight = TRUE)
@@ -30,4 +36,10 @@ test_that("error messages with wrong inputs", {
   expect_error(
     net_to_mat(net, missing_value = "zz"),
     "missing_value must be numeric.", fixed = TRUE)
+  
+  expect_error(
+    net_to_mat(net2, weight = TRUE),
+    "The third column of net must be numeric", fixed = TRUE)
+  
+  
 })