+++ NEWS.md
@@ -3,6 +3,9 @@
 This is a list of changes made in the development/github version of the package
 between bioregion 1.0.0 (CRAN release 2023-04-15) and the next CRAN release.
 
+* Added possibility to remove diagonal and lower triangular matrix in 
+mat_to_net() for squared matrix
+
 * Added a function to extract a subset of node according to its type (sites or 
 species) from a bioregion.clusters object containing both types of nodes (sites 
 and species).